+++ src/controllers/controllers.go
@@ -1,1 +1,3 @@
 package controllers
+
+func main() {}
