+++ src/routes/routes.go
@@ -2,7 +2,6 @@ package routes
 
 import (
 	"encoding/json"
-	"fmt"
 	"net/http"
 
 	"../models/"
@@ -59,16 +58,4 @@ func deleteReservation(w http.ResponseWriter, r *http.Request) {
 	}
 	delete(reservations, params["id"])
 	w.WriteHeader(http.StatusNoContent)
-}
-
-func main() {
-	api := mux.NewRouter()
-	api.handlefunc("/reservations", getReservations).Methods("GET")
-	api.handlefunc("/reservations/{id}", getReservation).Methods("GET")
-	api.HandleFunc("/users", createReservation).Methods("POST")
-	api.HandleFunc("/users/{id}", updateReservation).Methods("PUT")
-	api.HandleFunc("/users/{id}", deleteReservation).Methods("DELETE")
-
-	fmt.Printf("API listening on port 6000")
-	http.ListenAndServe(":6000", api)
 }
